refactor: adjust path
