+++ libs/json-difference/src/models/index.ts
@@ -1,1 +1,1 @@
-export * from './jsondiffer.model'
+export * from './json-difference.model'
